Add files via upload
This is a blackjack game through a console-based system made with Python. The rules of blackjack are as follows: Players receive all cards revealed and the robot's first card is face up and the second is face down. The object of the game is to get closer to 21 than the robot without going over 21. If a hand goes over 21, it is called a “bust” or “break” and you'll lose.